Use rocker/tidyverse as base image for integration tests
The rpy2 image was still using R v3.6.x, which was raising
weird errors when I tried to import fitzRoy. It was working fine
on my machine with Python v3.8.x and R v4.0.x, but rpy2 didn't
have an image with an updated R. There are a few extra steps to
install the correct versions of Python and R, but it works, and
that's what matters.